Move opplastningsboks.
